PRML filter and PRML classifier together with density estimation of mixing density.

PRML filter and PRML classifier together with density estimation of mixing density.

prml_tests_f(xA, xB, xAB, labels = c("A", "B", "AB"),
  remove.zeros = FALSE, mu_l = "min", mu_u = "max", e = 0,
  gamma.pars = c(0.5, 2e-10), n_gq = 20, n_mu = 100, n_per = 100,
  alpha = 0.5)

Arguments

xA
A vector. Spike counts of repeated dual-stimuli trial data AB.

xB
A vector. Spike counts of repeated single-stimulus trial data A.

xAB
A vector. Spike counts of repeated single-stimulus trial data B.

labels
A vector. labels for the trials.

remove.zeros
A logical value. Whether to remove 0s in spike counts.

mu_l
A number. Lower bound of spike counts. "min" by default. Indicating \( {max}(0,{min}{j=A,B,AB}({min}(Y_j)-2{std}(Y_j))) \)

mu_u
A number. Upper bound of spike counts. "max" by default. Indicating \( max_{j=A,B,AB}({max}(Y_j)+2{std}(Y_j)) \)

e
A number. 0 by default. Shringkage on the domain and meansurement of mixing density f under the Intermediate and Mixture hypothese.

gamma.pars
A length 2 vector. The shape and rate of gamma prior for spike rate mu_A and mu_B. Jeffereys' prior by default.

n_gq
A number. 20 by default. Number of grids in Gaussion quadrature.

n_mu
A number. 100 by default. The number of grids used to represent the pdf of f.

n_per
A number. 100 by default. Permutation of likihood estimation to obtain the order-invariant estimator.

alpha
0.5 by default. (For PRML filter) The range of the spike counts estimator \( [Y_{0.25}-\alpha {IQR},Y_{0.75}+\alpha {IQR}] \)

Value

A list.

out1

Result of prml_tests

out2

density estimation of mixing density f under Mixture, Intermediate, OutsideA, OutsideB hypotheses

See also